pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fede"
version = "0.1.0"
description = "Film Embedding and Discovery Engine — semantic and structural search over movie scripts"
requires-python = ">=3.12,<3.14"
dependencies = [
    # Numerics & data
    "numpy>=1.26",
    "pandas>=2.2,<3",
    # Deep learning & embeddings
    "torch>=2.5.1",
    "transformers>=4.46.3",
    "sentence-transformers>=3.2.1",
    "accelerate>=1.0.1",
    "datasets>=2.14.4",
    # Vector database
    "qdrant-client>=1.9.0",
    # Config & validation
    "typing-extensions>=4.11",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "nest-asyncio>=1.6.0",
    "google-genai>=1.0.0",
    # NLP
    "spacy>=3.7",
    "httpx>=0.27",
    # API layer
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "ipykernel>=6.31.0",
    "grafeo>=0.5.34,<0.6; python_version >= '3.12'",
    "peft>=0.17.1",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "httpx>=0.27",
    "jupyterlab>=4.1",
    "ipykernel>=6.29",
    "ruff>=0.4",
    "black>=22.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["vector_db", "knowledge_graph"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
